% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buf.R
\name{buf}
\alias{buf}
\title{Buffering introduced distributional records}
\usage{
buf(obs, b)
}
\arguments{
\item{obs}{occurrence records of longitude and latitude, input as geographic coordinate system (decimal degrees), WGS 1984}

\item{b}{distance for buffering, unit is meter}
}
\description{
Buffering introduced distributional records
}
