% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candD.R
\name{canD}
\alias{canD}
\title{Generate candidate patches/polygons from binary niche model predictions and/or buffered introduced distributional records}
\usage{
canD(pred2, obs, b)
}
\arguments{
\item{pred2}{ENM binary prediction}

\item{obs}{occurrence records of longitude and latitude, input as geographic coordinate system (decimal degrees), WGS 1984}

\item{b}{distance for buffering, unit is meter}
}
\description{
Generate candidate patches/polygons from binary niche model predictions and/or buffered introduced distributional records
}
