% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsAbs.R
\name{dsAbs}
\alias{dsAbs}
\title{Selecting pseudo-absence data for downscale coarse grain climate suitability prediction}
\usage{
dsAbs(sdm, tar, n)
}
\arguments{
\item{sdm}{ENM coarse grain climate suitability prediction rescale range from 0-1, whcih is also the extent where regional fine grain model will be calibrated.}

\item{tar}{fine grain environmental variables for regional habitat suitability modeling.}

\item{n}{number of pseudo-absence point to be selected.}
}
\description{
Selecting pseudo-absence data for downscale coarse grain climate suitability prediction
}
