% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmRoute.R
\name{enmRoute}
\alias{enmRoute}
\title{Survey_route_optimization}
\usage{
enmRoute(pred, p, r)
}
\arguments{
\item{pred}{ENM binary prediction}

\item{p}{Optimization by removing small size of pieces}

\item{r}{Optimization by remove low ranked patches}
}
\description{
Survey_route_optimization
}
