% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankCI.R
\name{rankCI}
\alias{rankCI}
\title{Ranking polygon patch, result of which used to determine parameter r in enmRoute}
\usage{
rankCI(pred, p)
}
\arguments{
\item{pred}{ENM binary prediction}

\item{p}{Optimization survey route by removing small size of pieces (km2)}
}
\description{
Ranking polygon patch, result of which used to determine parameter r in enmRoute
}
