% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thd.R
\name{thd}
\alias{thd}
\title{Thresholding model predictions}
\usage{
thd(sdm, points = NULL, type = NULL, threshold = NULL, binary = FALSE)
}
\arguments{
\item{sdm}{ENM suitability prediction}

\item{points}{Occ of long and lat}

\item{type}{mtp, p10}

\item{threshold}{a specific threshold value}

\item{binary}{whether to return the binary prediction}
}
\description{
Thresholding model predictions
}
