% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneSite.R
\name{tuneSite}
\alias{tuneSite}
\title{Tune sample site by estimating proportion of patches to be sampled
exploring relationship between sampled patches with driving time and accumulated capacity
driving time is related to distance between patches, surveying time in a specific patch is scaled with accumulated capacity}
\usage{
tuneSite(shp, r, u)
}
\arguments{
\item{shp}{is sf object, whcih is the output polygon of rankCI function in this package}

\item{r}{Optimization by removing low ranked patches, the number of patches to be removed}

\item{u}{increment of patches to be iteratively removed, values range between 1 and (dim(shp)[1]-5)}
}
\description{
Tune sample site by estimating proportion of patches to be sampled
exploring relationship between sampled patches with driving time and accumulated capacity
driving time is related to distance between patches, surveying time in a specific patch is scaled with accumulated capacity
}
