% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmRoute.R
\name{enmRoute}
\alias{enmRoute}
\title{Survey_route_optimization}
\usage{
enmRoute(pred1, pred2, p, r, obs, b)
}
\arguments{
\item{pred1}{Ecological Niche Model Suitability Prediction}

\item{pred2}{Ecological Niche Model Suitability Binary Prediction}

\item{p}{Optimization by removing small size of pieces}

\item{r}{Optimization by remove low ranked patches}

\item{obs}{occurrence records of longitude and latitude, input as geographic coordinate system (decimal degrees), WGS 1984}

\item{b}{distance for buffering, unit is kilometer}
}
\description{
Survey_route_optimization
}
