% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankCI.R
\name{rankCI}
\alias{rankCI}
\title{Ranking polygon patch by carrying capacity, result of which used to determine parameter r in enmRoute}
\usage{
rankCI(pred1, canD, p)
}
\arguments{
\item{pred1}{Ecological Niche Model Suitability Suitability Prediction, ranging 0-1000}

\item{canD}{candidate polygons of patches for optimization}

\item{p}{small size of pieces (km2) removing before optimization}
}
\description{
Ranking polygon patch by carrying capacity, result of which used to determine parameter r in enmRoute
}
