% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thd.R
\name{thd}
\alias{thd}
\title{Thresholding ecological niche model suitability predictions}
\usage{
thd(pred1, points = NULL, type = NULL, threshold = NULL, binary = FALSE)
}
\arguments{
\item{pred1}{ENM suitability prediction, ranging 0-1000.}

\item{points}{obs occurrence records of longitude and latitude, input as geographic coordinate system (decimal degrees), WGS 1984}

\item{type}{mtp, p10}

\item{threshold}{a specific threshold value}

\item{binary}{whether to return the binary prediction}
}
\description{
Thresholding ecological niche model suitability predictions
}
